// client/src/app/_models/member.ts
import { Photo } from "./Photo"

export interface Member {
  id: number
  username: string
  age: number
  photoUrl: string
  knownAs: string
  created: Date
  lastActive: Date
  gender: string
  introduction: string
  interests: string
  lookingFor: any
  city: string
  country: string
  photos: Photo[]
}